proto: support protobuf any
This commit adds better support for protobuf deserialization by
allowing the console to deserialize custom proto messages that
are supplied to fields with the `google.protobuf.Any` type.

Before this change, if your protobuf message
contained a field of type `google.protobuf.Any` and the value
supplied at runtime was not a standard proto message
(e.g. `google.protobuf.Timestamp`), then `msg.MarshalJSON`
(called from `deserializeProtobufMessageToJSON`) would yield
the following error:

```
failed to marshal protobuf message to JSON:
unknown message type \"<type-url>\"
```

This error was emitted by the `github.com/jhump/protoreflect` because
it would fail to resolve the 'type url' of the value supplied to the
`google.protobuf.Any` field. This error would
occur even if the type was available in the proto registry.

The error occurred, because the internal `AnyResolver` created by
`github.com/jhump/protoreflect` only contained the file descriptor
relating to the overall protobuf message being processed.
Therefore if the type referred to by the
'google.protobuf.Any' field existed in some other file,
the type could not be resolved.

This commit amends this behaviour by calling
`msg.MarshalJSONPB` and passing a custom `AnyResolver` through,
defined in `console/backend/pkg/proto/service.go`.
The custom `AnyResolver` takes precedence over the internal
`AnyResolver`, but the internal `AnyResolver` is still used by
`github.com/jhump/protoreflect` if the custom `AnyResolver` cannot
resolve a type. The custom `AnyResolver` provided,
strips the "type url" of anything before the last '/' and then calls
the proto registry to resolve the type. Removing the contents of the
string before the slash is necessary because types begin with
"type.googleapis.com/", but the proto registry does not contain this
prefix.

The result of the change is that registry types can now be considered,
which allows the console to support deserialization of custom proto
message types that are supplied to a `google.protobuf.Any` field.